+++ Sources/StreamChat/Workers/ChannelUpdater.swift
@@ -99,10 +99,25 @@ class ChannelUpdater: Worker {
 
     /// Updates specific channel with new data.
     /// - Parameters:
-    ///   - channelPayload: New channel data..
+    ///   - channelPayload: New channel data.
     ///   - completion: Called when the API call is finished. Called with `Error` if the remote update fails.
     func updateChannel(channelPayload: ChannelEditDetailPayload, completion: ((Error?) -> Void)? = nil) {
         apiClient.request(endpoint: .updateChannel(channelPayload: channelPayload)) {
+            completion?($0.error)
+        }
+    }
+
+    /// Updates specific channel with provided data, and removes unneeded properties.
+    /// - Parameters:
+    ///   - updates: Updated channel data. Only non-nil data will be updated.
+    ///   - unsetProperties: Properties from the channel that are going to be cleared/unset.
+    ///   - completion: Called when the API call is finished. Called with `Error` if the remote update fails.
+    func partialChannelUpdate(
+        updates: ChannelEditDetailPayload,
+        unsetProperties: [String],
+        completion: ((Error?) -> Void)? = nil
+    ) {
+        apiClient.request(endpoint: .partialChannelUpdate(updates: updates, unsetProperties: unsetProperties)) {
             completion?($0.error)
         }
     }